PhysiCellCellCreator

Documentation for PhysiCellCellCreator. A tool to create initial conditions for PhysiCell simulations.

PhysiCellCellCreator.createICCellXMLTemplate

PhysiCellCellCreator.createICCellXMLTemplate — Method

createICCellXMLTemplate(folder::String)

Create folder path_to_folder and create a template XML file for IC cells.

Introduces a new way to initialize cells in a simulation, wholly contained within a Julia package and outside a PhysiCell runtime. It will not work in PhysiCell! This function creates a template XML file for IC cells, showing all the current functionality of this initialization scheme. It uses the cell type "default". Create ICs for more cell types by copying the cell_patches element.